Rename model to cli.
